+++ app/src/main/java/woowacourse/shopping/presentation/view/orderlist/OrderListActivity.kt
@@ -12,6 +12,7 @@ import woowacourse.shopping.presentation.model.OrderDetailModel
 import woowacourse.shopping.presentation.view.orderdetail.OrderDetailActivity
 import woowacourse.shopping.presentation.view.orderlist.adapter.OrderListAdapter
 import woowacourse.shopping.presentation.view.productlist.ProductListActivity.Companion.KEY_SERVER_SERVER
+import woowacourse.shopping.presentation.view.util.createRetrofit
 import woowacourse.shopping.presentation.view.util.getSerializableCompat
 import woowacourse.shopping.presentation.view.util.showToast
 
@@ -33,7 +34,8 @@ class OrderListActivity : AppCompatActivity(), OrderListContract.View {
 
     private fun setPresenter() {
         server = intent.getSerializableCompat(KEY_SERVER_SERVER) ?: return finish()
-        val orderRepository = OrderRepositoryImpl(server)
+        val retrofit = createRetrofit(server)
+        val orderRepository = OrderRepositoryImpl(retrofit)
         presenter = OrderListPresenter(this, orderRepository)
     }
